another refactor pass
