+++ api/prism_app/tests/test_export_s3.py
@@ -4,22 +4,15 @@ from unittest.mock import patch
 import pytest
 from prism_app.export_s3 import (
     DEFAULT_EXPORT_MAP_S3_BUCKET,
-    MAP_EXPORT_S3_SIGNING_REGION,
-    get_export_map_s3_bucket_and_prefix,
     is_file_artifact_uri,
     local_path_from_file_uri,
     map_export_s3_client,
     normalize_export_map_s3_object_prefix,
     parse_export_map_s3_bucket_env,
     parse_s3_uri,
-    public_maps_folder_prefix,
-    public_maps_folder_uri,
-    public_maps_root_admin_output_path,
-    public_maps_root_folder_uri,
     put_map_export_bytes_local,
     s3_key_for_map_export,
     slug_s3_path_segment,
-    storage_uri_to_admin_output_path,
 )
 
 
@@ -69,97 +62,6 @@ def test_s3_key_for_map_export_public_maps():
         )
         == "public_maps/mozambique/precip_blended_dekad/jid.pdf"
     )
-    export_url = (
-        "https://prism.example/export?date={date}"
-        "&hazardLayerIds=precip_blended_dekad"
-    )
-    assert (
-        public_maps_folder_prefix(export_url, country="Mozambique")
-        == "public_maps/mozambique/precip_blended_dekad/"
-    )
-
-
-def test_public_maps_folder_uri_includes_bucket_and_prefix(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    monkeypatch.delenv("EXPORT_MAP_LOCAL_OUTPUT_DIR", raising=False)
-    monkeypatch.delenv("EXPORT_MAP_S3_BUCKET", raising=False)
-    export_url = (
-        "https://prism.example/export?date={date}"
-        "&hazardLayerIds=precip_blended_dekad"
-    )
-    assert public_maps_folder_uri(export_url, country="Mozambique") == (
-        "s3://prism-wfp/batch-maps/public_maps/mozambique/precip_blended_dekad/"
-    )
-    assert get_export_map_s3_bucket_and_prefix() == ("prism-wfp", "batch-maps")
-
-
-def test_public_maps_root_folder_uri_s3(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    monkeypatch.delenv("EXPORT_MAP_LOCAL_OUTPUT_DIR", raising=False)
-    monkeypatch.delenv("EXPORT_MAP_S3_BUCKET", raising=False)
-    assert public_maps_root_folder_uri() == ("s3://prism-wfp/batch-maps/public_maps/")
-    link = public_maps_root_admin_output_path()
-    assert "s3.console.aws.amazon.com/s3/buckets/prism-wfp" in link
-    assert "prefix=batch-maps/public_maps/" in link
-
-
-def test_public_maps_root_folder_uri_local(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    monkeypatch.setenv("EXPORT_MAP_S3_BUCKET", "")
-    monkeypatch.setenv("EXPORT_MAP_LOCAL_OUTPUT_DIR", str(tmp_path))
-    assert public_maps_root_folder_uri().startswith("file://")
-    path = public_maps_root_admin_output_path()
-    assert path == str((tmp_path / "public_maps").resolve())
-    assert not path.startswith("file:")
-
-
-def test_storage_uri_to_admin_output_path_s3_console(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    monkeypatch.delenv("EXPORT_MAP_LOCAL_OUTPUT_DIR", raising=False)
-    storage_uri = (
-        "s3://prism-wfp/batch-maps/public_maps/mozambique/precip_blended_dekad/"
-    )
-    link = storage_uri_to_admin_output_path(storage_uri)
-    assert link.startswith("https://s3.console.aws.amazon.com/s3/buckets/prism-wfp")
-    assert f"region={MAP_EXPORT_S3_SIGNING_REGION}" in link
-    assert "prefix=batch-maps/public_maps/mozambique/precip_blended_dekad/" in link
-
-
-def test_storage_uri_to_admin_output_path_local_file_scheme(
-    tmp_path: Path,
-) -> None:
-    folder = tmp_path / "public_maps" / "moz" / "layer"
-    storage_uri = folder.as_uri() + "/"
-    assert storage_uri_to_admin_output_path(storage_uri) == str(folder) + "/"
-
-
-def test_storage_uri_to_admin_output_path_bare_prefix_uses_bucket(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    monkeypatch.delenv("EXPORT_MAP_LOCAL_OUTPUT_DIR", raising=False)
-    monkeypatch.delenv("EXPORT_MAP_S3_BUCKET", raising=False)
-    link = storage_uri_to_admin_output_path(
-        "batch-maps/public_maps/mozambique/precip_blended_dekad/",
-    )
-    assert "s3.console.aws.amazon.com/s3/buckets/prism-wfp" in link
-    assert "prefix=batch-maps/public_maps/mozambique/precip_blended_dekad/" in link
-
-
-def test_public_maps_folder_uri_local_file_scheme(
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-) -> None:
-    monkeypatch.setenv("EXPORT_MAP_S3_BUCKET", "")
-    monkeypatch.setenv("EXPORT_MAP_LOCAL_OUTPUT_DIR", str(tmp_path))
-    export_url = "https://prism.example/export?date={date}&hazardLayerIds=my_layer"
-    uri = public_maps_folder_uri(export_url, country="cambodia")
-    assert uri.startswith("file://")
-    assert "public_maps/cambodia/my_layer" in uri
     assert (
         s3_key_for_map_export(
             "jid",
